added tools.py; added more tests; attempted to change main.py back to using ThreadPool but api is now broken
