+++ api/tests/test_api.py
@@ -7,7 +7,7 @@ def test_ping(client):
     assert output.status_code == 200
 
 
-def test_single_tag(client):
+def test_single_tag_sorted(client):
     """Verify single tag with sorting"""
     output = client.get(
         "http://localhost:5000/posts?tag=history&sortBy=likes&direction=desc"
@@ -20,7 +20,7 @@ def test_single_tag(client):
     assert output == expected
 
 
-def test_multi_tag(client):
+def test_multi_tag_sorted(client):
     """Verify multi tag with sorting"""
     output = client.get(
         "http://localhost:5000/posts?tag=history,tech&sortBy=likes&direction=desc"
@@ -33,7 +33,7 @@ def test_multi_tag(client):
     assert output == expected
 
 
-def test_multi_tag2(client):
+def test_multi_tag2_sorted(client):
     """Verify multi tag with sorting"""
     output = client.get(
         "http://localhost:5000/posts?tag=politics,art&sortBy=likes&direction=desc"
@@ -46,10 +46,23 @@ def test_multi_tag2(client):
     assert output == expected
 
 
-def no_tags(client):
-    """verify that we error out correctly when no tags applied"""
-    output = client.get("http://localhost:5000/posts").get_json()
+def test_single_tag(client):
+    """Verify single tag with sorting"""
+    output = client.get("http://localhost:5000/posts?tag=history").get_json()
 
-    expected = requests.get("https://api.hatchways.io/assessment/solution/posts").json()
+    expected = requests.get(
+        "https://api.hatchways.io/assessment/solution/posts?tags=history"
+    ).json()
+
+    assert output == expected
+
+
+def test_multi_tag(client):
+    """Verify multi tag with sorting"""
+    output = client.get("http://localhost:5000/posts?tag=history,tech").get_json()
+
+    expected = requests.get(
+        "https://api.hatchways.io/assessment/solution/posts?tags=history,tech"
+    ).json()
 
     assert output == expected
